refactor(samples): split query samples into separate files
